workspace conditions
