+++ src/packages/core/extension-registry/conditions/types.ts
@@ -4,11 +4,6 @@ import type { UmbSectionUserPermissionConditionConfig } from '../../section/cond
 import type { SectionAliasConditionConfig } from './section-alias.condition.js';
 import type { SwitchConditionConfig } from './switch.condition.js';
 import type { UmbMultipleAppLanguageConditionConfig } from './multiple-app-languages.condition.js';
-import type {
-	WorkspaceAliasConditionConfig,
-	WorkspaceEntityTypeConditionConfig,
-	WorkspaceContentTypeAliasConditionConfig,
-} from '@umbraco-cms/backoffice/workspace';
 import type { UmbConditionConfigBase } from '@umbraco-cms/backoffice/extension-api';
 import type { UmbDocumentUserPermissionConditionConfig } from '@umbraco-cms/backoffice/document';
 
@@ -29,10 +24,7 @@ export type ConditionTypes =
 	| UmbConditionConfigBase
 	| UmbDocumentUserPermissionConditionConfig
 	| UmbMultipleAppLanguageConditionConfig
-	| UmbSectionUserPermissionConditionConfig
-	| WorkspaceAliasConditionConfig
-	| WorkspaceContentTypeAliasConditionConfig
-	| WorkspaceEntityTypeConditionConfig;
+	| UmbSectionUserPermissionConditionConfig;
 
 type UnionOfProperties<T> = T extends object ? T[keyof T] : never;
