+++ src/packages/core/workspace/conditions/const.ts
@@ -1,0 +1,1 @@
+export const UMB_WORKSPACE_HAS_COLLECTION_CONDITION = 'Umb.Condition.WorkspaceHasCollection';
