+++ src/packages/core/workspace/conditions/index.ts
@@ -1,5 +1,1 @@
-export type { WorkspaceAliasConditionConfig } from './workspace-alias.condition.js';
-export type { WorkspaceEntityTypeConditionConfig } from './workspace-entity-type.condition.js';
-export type { WorkspaceHasCollectionConditionConfig } from './workspace-has-collection.condition.js';
-export type { WorkspaceContentTypeAliasConditionConfig as WorkspaceEntityContentTypeConditionConfig } from './workspace-content-type-alias.condition.js';
-export type { WorkspaceContentTypeAliasConditionConfig } from './workspace-content-type-alias.condition.js';
+export type * from './types.js';
