+++ src/packages/core/workspace/conditions/workspace-entity-type.condition.ts
@@ -1,8 +1,8 @@
 import { UMB_WORKSPACE_CONTEXT } from '../contexts/index.js';
+import type { WorkspaceEntityTypeConditionConfig } from './types.js';
 import { UmbConditionBase } from '@umbraco-cms/backoffice/extension-registry';
 import type {
 	ManifestCondition,
-	UmbConditionConfigBase,
 	UmbConditionControllerArguments,
 	UmbExtensionCondition,
 } from '@umbraco-cms/backoffice/extension-api';
@@ -19,15 +19,6 @@ export class UmbWorkspaceEntityTypeCondition
 		});
 	}
 }
-
-export type WorkspaceEntityTypeConditionConfig = UmbConditionConfigBase<'Umb.Condition.WorkspaceEntityType'> & {
-	/**
-	 * Define the workspace that this extension should be available in
-	 * @example
-	 * "Document"
-	 */
-	match: string;
-};
 
 export const manifest: ManifestCondition = {
 	type: 'condition',
