+++ src/packages/core/workspace/conditions/workspace-has-collection.condition.ts
@@ -1,12 +1,15 @@
 import { UMB_COLLECTION_WORKSPACE_CONTEXT } from '../contexts/tokens/collection-workspace.context-token.js';
+import type { WorkspaceHasCollectionConditionConfig } from './types.js';
+import { UMB_WORKSPACE_HAS_COLLECTION_CONDITION } from './const.js';
 import { UmbConditionBase } from '@umbraco-cms/backoffice/extension-registry';
 import type {
 	ManifestCondition,
-	UmbConditionConfigBase,
 	UmbConditionControllerArguments,
 	UmbExtensionCondition,
 } from '@umbraco-cms/backoffice/extension-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
+
+const ObserveSymbol = Symbol();
 
 export class UmbWorkspaceHasCollectionCondition
 	extends UmbConditionBase<WorkspaceHasCollectionConditionConfig>
@@ -21,17 +24,11 @@ export class UmbWorkspaceHasCollectionCondition
 				(hasCollection) => {
 					this.permitted = hasCollection;
 				},
-				'observeHasCollection',
+				ObserveSymbol,
 			);
 		});
 	}
 }
-
-export type WorkspaceHasCollectionConditionConfig = UmbConditionConfigBase<
-	typeof UMB_WORKSPACE_HAS_COLLECTION_CONDITION
->;
-
-export const UMB_WORKSPACE_HAS_COLLECTION_CONDITION = 'Umb.Condition.WorkspaceHasCollection';
 
 export const manifest: ManifestCondition = {
 	type: 'condition',
